test contex -> text contexts
